refact: 컴포넌트화 리팩토링
