+++ src/App.tsx
@@ -4,24 +4,24 @@ import Community from '@/pages/community/Community';
 import FriendList from '@/pages/friendList/FriendList';
 import Setting from '@/pages/settings/Setting';
 import NavBar from '@/components/Navbar';
-import ChattingRoom from '@/pages/chat/ChattingRoom';
+import ChattingRoom from '@/pages/chat/chattingRoom/ChattingRoom';
 
 import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom';
-import Profile from '@/pages/friendList/Profile';
+import Profile from '@/pages/friendList/profile/Profile';
 import Statusbar from '@/components/Statusbar';
-import FriendProfile from '@/pages/friendList/FriendProfile';
+import FriendProfile from '@/pages/friendList/profile/FriendProfile';
 
 function AppLayout({ children }: { children: React.ReactNode }) {
   const location = useLocation();
   const hideSidebarPaths = ['/chattingroom', '/profile']; // 숨기고 싶은 페이지 경로
-  const showSidebar = !hideSidebarPaths.some((path) => location.pathname.startsWith(path)); //pathname이 /myprofile/123 이런식이어도 적용됨
+  const showNavbar = !hideSidebarPaths.some((path) => location.pathname.startsWith(path)); //pathname이 /myprofile/123 이런식이어도 적용됨
   const isChatRoom = location.pathname.startsWith('/chattingroom');
   const isProfile = location.pathname.startsWith('/profile');
 
   return (
     <div className="relative">
       <Statusbar isChatRoom={isChatRoom} isProfile={isProfile} />
-      {showSidebar && <NavBar />}
+      {showNavbar && <NavBar />}
       <div className="min-h-screen">{children}</div>
     </div>
   );
@@ -43,7 +43,6 @@ export default function App() {
           <Route path="/profile/:id" element={<FriendProfile />} />
         </Routes>
       </AppLayout>
-      {/*<Sidebar />  하단 고정 */}
     </Router>
   );
 }
