+++ src/pages/friendList/profile/FriendProfile.tsx
